LLM/requirements.txt
fastapi==0.111.0
nltk==3.8.1
openai==1.35.13
python-dotenv==1.0.1
pywin32==306
